requirements.txt
python-binance>=1.0.19
typer>=0.12
rich>=13.7
python-dotenv>=1.0
